Move some interfaces from admin back to consumer.
